fix image url

+++ app/orders/views.py
@@ -1,9 +1,10 @@
 from django.shortcuts import render
 from .models import Order, OrderChecking
 from index.models import User
+from customer.models import CustomerProfile
 from .serializers import OrderSerializer
 from rest_framework.views import APIView
-from rest_framework.permissions import IsAuthenticated
+from rest_framework.permissions import IsAuthenticated, IsAdminUser
 from rest_framework.parsers import MultiPartParser, FormParser
 from drf_yasg.utils import swagger_auto_schema
 from drf_yasg import openapi
@@ -18,6 +19,8 @@ from datetime import timedelta
 from utilities.models import SMS
 from utilities.sms import send_sms
 from utilities.mapping import send_confirm_sms
+
+
 # Create your views here.
 
 
@@ -67,23 +70,35 @@ class OrderAPIView(APIView):
     @swagger_auto_schema(request_body=OrderSerializer, parser_classes=parser_classes,
                          manual_parameters=[order_id, param_config])
     def put(self, request):
-        order_id = request.data["order_id"]
-        order = Order.objects.get(id=int(order_id))
-        serializer = OrderSerializer(order, many=False, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_200_OK)
+        customer_id = request.data["customer"]
+        customer = CustomerProfile.get(id=customer_id)
+        user_type = customer.user.user_type
+        if user_type <= 2:
+            order_id = request.data["order_id"]
+            order = Order.objects.get(id=int(order_id))
+            serializer = OrderSerializer(order, many=False, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            else:
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response("not accessible for customer and user", status=status.HTTP_400)
 
     @swagger_auto_schema(manual_parameters=[order_id, param_config])
     def delete(self, request):
-        try:
-            order = Order.objects.get(id=int(request.data['order_id']))
-            order.delete()
-            return Response({'detail': 'Order is deleted'}, status=status.HTTP_200_OK)
-        except:
-            return Response({'detail': 'Order does not exist'}, status=status.HTTP_400_BAD_REQUEST)
+        customer_id = request.data["customer"]
+        customer = CustomerProfile.get(id=customer_id)
+        user_type = customer.user.user_type
+        if user_type <= 2:
+            try:
+                order = Order.objects.get(id=int(request.data['order_id']))
+                order.delete()
+                return Response({'detail': 'Order is deleted'}, status=status.HTTP_200_OK)
+            except:
+                return Response({'detail': 'Order does not exist'}, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return Response("not accessible customer and user", status=status.HTTP_400_BAD_REQUEST)
 
     @swagger_auto_schema(manual_parameters=[param_config])
     def get(self, request):
@@ -91,19 +106,58 @@ class OrderAPIView(APIView):
         serializer = OrderSerializer(order, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-# class SingleBrandAPI(APIView):
-#     permission_classes = [IsAuthenticated, IsAdminUser]
-#     serializer_class = BrandSerializer
+
+class SingleOrderAPI(APIView):
+    permission_classes = [IsAuthenticated]
+    serializer_class = OrderSerializer
+    parser_classes = (MultiPartParser, FormParser)
+
+    token = openapi.Parameter('Authorization', in_=openapi.IN_HEADER,
+                              description='enter access token', type=openapi.TYPE_STRING, )
+
+    @swagger_auto_schema(manual_parameters=[token], parser_classes=parser_classes)
+    def get(self, request, pk):
+        try:
+            order = Order.objects.get(id=pk)
+            serializer = OrderSerializer(order, many=False)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        except:
+            return Response({'detail': 'Order does not exist'}, status=status.HTTP_400_BAD_REQUEST)
+
+
+# class OrderCancelAPI(APIView):
 #     parser_classes = (MultiPartParser, FormParser)
-#
-#     token = openapi.Parameter('Authorization', in_=openapi.IN_HEADER,
-#                               description='enter access token', type=openapi.TYPE_STRING, )
-#
-#     @swagger_auto_schema(manual_parameters=[token], parser_classes=parser_classes)
-#     def get(self, request, pk):
-#         try:
-#             brand = Brand.objects.get(id=pk)
-#             serializer = BrandSerializer(brand, many=False, context={"request": request})
-#             return Response(serializer.data)
-#         except:
-#             return Response({'detail': 'Brand does not exist'})
+
+    # @swagger_auto_schema(request_body=openapi.Schema(
+    #     type=openapi.TYPE_OBJECT,
+    #     properties={
+    #         'user_id': openapi.Schema(type=openapi.TYPE_STRING, description='The desc'),
+    #         'order_id': openapi.Schema(type=openapi.TYPE_STRING, description='The desc'),
+    #     }
+    # ))
+    # user_id = openapi.Parameter(
+    #     'user_id',
+    #     in_=openapi.IN_QUERY,
+    #     description='Enter user id to verify the user ',
+    #     type=openapi.TYPE_INTEGER
+    # )
+    # order_id = openapi.Parameter(
+    #     'order id',
+    #     in_=openapi.IN_QUERY,
+    #     description='Enter order id the user ',
+    #     type=openapi.TYPE_INTEGER
+    # )
+    #
+    # @swagger_auto_schema(manual_parameters=[user_id, order_id])
+    # def post(self, request):
+    #     order_id = request.GET.get('order_id')
+    #     user_id = request.GET.get('user_id')
+    #     print(order_id)
+    #     # order_id = data["order_id"]
+    #     # user_id = data["user_id"]
+    #     order_Checking_itself = OrderChecking.objects.get(order_id=order_id)
+    #     print(order_Checking_itself)
+    #     order_Checking_itself.confirmed = 2
+    #     order_Checking_itself.save()
+    #     print(order_Checking_itself)
+    #     return Response("order cancel chenged")
